qbism bsp loading works - both collision and drawing
todo:
- cm_world is static alloated so loading anything over vanilla limits will crash it
- PVS/PHS needs adjusting (limited to vanilla sizes)

+++ src/renderer_gl2/r_model.h
@@ -55,11 +55,10 @@ typedef struct
 #define SURF_DRAWBACKGROUND	0x40
 #define SURF_UNDERWATER		0x80
 
-// !!! if this is changed, it must be changed in asm_draw.h too !!!
 typedef struct
 {
-	unsigned short	v[2];
-	unsigned int	cachededgeoffset;
+	unsigned int	v[2]; // braxi -- was unsigned short
+	unsigned int	cachededgeoffset; // braxi -- unused
 } medge_t;
 
 typedef struct mtexinfo_s
@@ -129,8 +128,8 @@ typedef struct mnode_s
 	cplane_t	*plane;
 	struct mnode_s	*children[2];	
 
-	unsigned short		firstsurface;
-	unsigned short		numsurfaces;
+	unsigned int		firstsurface; // braxi -- was unsigned short
+	unsigned int		numsurfaces; // braxi -- was unsigned short
 } mnode_t;
